Fix duplicate code in Difficulty.cs (due to bad merge) Fixed type for Song.Sections Added "this" code style rule to .editorconfig

+++ MediumHardMod/Song.cs
@@ -15,7 +15,7 @@ internal class Song
     public decimal Offset { get; set; }
     public decimal Length { get; set; }
     public int BeatsPerMeasure { get; set; }
-    public Dictionary<int, string> Sections { get; set; } = new();
+    public Dictionary<double, string> Sections { get; set; } = new();
     public List<SongChart> SongCharts { get; set; } = new();
 
     public bool HasChart(Difficulty difficulty)
